Director of Impact: Part of business AC. Leads all Impact efforts including documentation, presentation, and scheduling practices. Presents Impact alongside two other presenters chosen via tryouts.
Impact is an integral part of our team. The Impact lead deserves to be recognized for their hard work and considered a leader of the team. They deserve to have a say in matters the AC decides on. By granting them a say in decision-making processes and including them in relevant channels, we not only acknowledge their contributions but also empower them to further impact the team's direction and outcomes. Our team is also more likely to take Impact seriously if there's a position for it.