EC should be required to attend all competitions/regionals during the season, unless extenuating circumstances (health, family emergencies, etc.) which must be explained to Mr. Blay ASAP.
As a robotics team, competitions are the ultimate showcase of our success and achievements. Winning awards, participating in outreach events, and interacting with teams from around the world not only validate our hard work but also provide invaluable learning experiences and networking opportunities. For this reason, it is imperative that every leader of our team commits to attending all competitions and regionals. These events are where our team's dedication, innovation, and teamwork are put to the test, and the presence of our leaders is vital in representing our collective efforts. Furthermore, if a leader of the team doesn't attend every competition, who will? Being a leader means you have made a commitment to the team and are there to support the team in both good and bad times. To be a good leader is to be engaged with the team, and to do that, you must attend every competition where our team could possibly reach its peak (or a very sad valley). Overall, requiring EC members to attend all competitions and regionals ensures effective leadership, strengthens team cohesion, and maximizes the team's opportunities for success both on and off the field.